% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_longlat}
\alias{all_longlat}
\title{All items in list in Long Lat coordinateS?}
\usage{
all_longlat(pops)
}
\arguments{
\item{pops}{list of rasters}
}
\value{
boolean
}
\description{
All items in list in Long Lat coordinateS?
}
