% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoboundaries}
\alias{geoboundaries}
\title{Data frame of available datasets through the geoBoundaries API.}
\format{
A data frame with 644 rows and 6 variables:
\describe{
\item{iso_code}{the three letter iso code for each country, see the source url for more information}
\item{year}{the year of the dataset}
\item{admin_level}{the admin level of the available dataset}
\item{source}{the source of the dataset}
\item{license}{the license of the dataset}
\item{country}{the country name}
}
}
\source{
\url{https://www.geoboundaries.org/api.html}
}
\usage{
geoboundaries
}
\description{
Data frame of available datasets through the geoBoundaries API.
}
\keyword{dataset}
