% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_country_shape}
\alias{get_country_shape}
\title{Get country shapefile from geoboundaries}
\usage{
get_country_shape(country_iso, admin_level, type = "SSCU")
}
\arguments{
\item{country_iso}{the three letter code corresponding to the country.
Use ?popcompr::iso_codes to see the full list.}

\item{admin_level}{numeric, the admin level of the shapefile (0 = the country border,
then 1, 2, 3). Note that not all admin levels are available}

\item{type}{the type of shapefile. Geoboundaries have simplified and unsimplified,
and standardized vs. unstandardized (and combinations thereof) available
for use. See the api documentation for more details: \url{https://www.geoboundaries.org/api.html}}
}
\value{
a shapefile as an sf object
}
\description{
Get country shapefile from geoboundaries
}
\examples{
# Defaults to simplified
les_simple <- get_country_shape(country_iso = "LSO", admin_level = 2)

}
