% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iso_codes}
\alias{iso_codes}
\title{List of country iso codes.}
\format{
A data frame with 246 rows and 2 variables:
\describe{
\item{iso_code}{the three letter iso code for each country, see the source url for more information}
\item{country}{the country name}
}
}
\source{
\url{https://unstats.un.org/unsd/tradekb/knowledgebase/country-code}
}
\usage{
iso_codes
}
\description{
List of country iso codes.
}
\keyword{datasets}
