% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lesotho_fb_2019}
\alias{lesotho_fb_2019}
\title{Population counts for Lesotho from Facebook/CIESIN for the year 2019.}
\format{
A geotiff file with number of people per pixel estimated by Facebook/CIESIN
for the year 2020 at a resolution of 3 arc seconds (apprx 100m at the equator),
projected in Geographic Coordinate System, WGS84.
See source url for more details.
}
\source{
\url{https://data.humdata.org/dataset/highresolutionpopulationdensitymaps-lso}
}
\usage{
lesotho_fb_2019
}
\description{
Population counts for Lesotho from Facebook/CIESIN for the year 2019.
}
\keyword{datasets}
