% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lso_facebook_2019}
\alias{lso_facebook_2019}
\title{Population counts for Lesotho from Facebook/CIESIN for the year 2019.}
\format{
A geotiff file with number of people per pixel estimated by Facebook/CIESIN
for the year 2020 at a resolution of 3 arc seconds (apprx 100m at the equator),
projected in Geographic Coordinate System, WGS84.
See source url for more details.
}
\source{
\url{https://data.humdata.org/dataset/highresolutionpopulationdensitymaps-lso}
}
\usage{
lso_facebook_2019
}
\description{
Population counts for Lesotho from Facebook/CIESIN for the year 2019.
}
\details{
This is just a help file, to access the file see the example below.
}
\examples{
\dontrun{
lesotho_fb_2019 <- raster(system.file("external/lso_facebook_2019.tif", package="popcompr"))
}


}
\keyword{dataset}
