% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{match_nearest}
\alias{match_nearest}
\title{Match to closest nonNA}
\usage{
match_nearest(cell_ids, to_match, max_adjacent = 100)
}
\arguments{
\item{cell_ids}{NA cells}

\item{to_match}{fasterized raster}

\item{max_adjacent}{max window to look out}
}
\description{
Matches cells in pop raster which do not have a match to the friction surface
to nearest cell that does have a match to the friction surface. Mainly issue with coastal
populations.
}
\keyword{interal}
