% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_compare}
\alias{plot_compare}
\title{Autoplot functions for pop comparisons}
\usage{
plot_compare(
  comp_obj,
  select_pops = NULL,
  type = "map",
  interactive = FALSE,
  save = FALSE,
  path = NULL,
  ext = "jpeg"
)
}
\arguments{
\item{comp_obj}{sf or RasterBrick object generated by comparison functions (see Details)}

\item{select_pops}{a numeric or character vector of either numeric indices or
colum names, defaults to all population columns (i.e. ones with the prefix "popcmp")}

\item{type}{character with option of the type of plots (all ggplots):
"hex" (requires \code{hexbin}), "pairs" (requires \code{GGally}),
"map" (requires and \code{sf}), and "hist". Defaults to "map".}

\item{interactive}{logical, should the plot be returned as an interactive
html plot using \code{plotly}? Defaults to FALSE.}

\item{save}{logical, should the plot be saved rather than returned? Requires
both path & ext params.}

\item{path}{if save = TRUE, then the filepath to write the plot out to}

\item{ext}{file extension (i.e. "jpeg", "png", "tiff")}
}
\value{
either a ggplot or a message stating where the ggplot was written
out to.
}
\description{
This function generates automatic comparison plots of pixel level
or admin level population comparisons from the \code{\link{compare_pop}}
(a raster brick), or the \code{\link{aggregate_to_shp}} (an sf object). If
comparing a large number of rasters, you should be careful of memory
requirements (a general guideline is that ncell(exe) * nlayers(exe) should not
exceed 1e8, but this may be system dependent).
}
