% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{resample_fun}
\alias{resample_fun}
\title{Title}
\usage{
resample_fun(pop, template, parallel, estimate_time)
}
\arguments{
\item{pop}{pop raster}

\item{template}{template to resample to}

\item{parallel}{run parallel}

\item{estimate_time}{}
}
\value{
a data.table or time estimate
}
\description{
Title
}
\keyword{internal}
