% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{resample_to_template}
\alias{resample_to_template}
\title{Title}
\usage{
resample_to_template(pop, template, parallel, estimate_time)
}
\arguments{
\item{pop}{pop raster}

\item{template}{template to resample to}

\item{parallel}{run parallel}

\item{estimate_time}{estimate time?}
}
\description{
Title
}
\keyword{internal}
