% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summarize_pops}
\alias{summarize_pops}
\title{Summary function for pop comparisons}
\usage{
summarize_pops(
  comp_obj,
  select_pops = NULL,
  funs = list(sum = function(x) sum(x, na.rm = TRUE), mean = function(x) mean(x, na.rm
    = TRUE), max = function(x) max(x, na.rm = TRUE), min = function(x) min(x, na.rm =
    TRUE), sd = function(x) sd(x, na.rm = TRUE), length_nas = function(x) sum(is.na(x)))
)
}
\arguments{
\item{comp_obj}{sf or RasterBrick object generated by comparison functions (see Details)}

\item{select_pops}{a numeric or character vector of either numeric indices or
colum names, defaults to all population columns (i.e. ones with the prefix "popcmp")}

\item{funs}{list of functions, you can pass your own with a single parameter
x or you can use the defaults (sum, mean, max, min, sd, and length_nas)}
}
\value{
a table with the rows corresponding to the summary stats and the
columns the population datasets
}
\description{
This function generates automatic summary stats of pixel level
or admin level population comparisons from the \code{\link{compare_pop}}
(a raster brick), or the \code{\link{aggregate_to_shp}} (an sf object).
}
