% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_to_shp}
\alias{aggregate_to_shp}
\title{Title}
\usage{
aggregate_to_shp(brick, sf, max_adjacent = 100)
}
\arguments{
\item{brick}{brick of pops}

\item{sf}{sf object}

\item{max_adjacent}{window to look at when matching to non Nas}
}
\value{

}
\description{
Title
}
