% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_pop}
\alias{compare_pop}
\title{Title}
\usage{
compare_pop(
  pops,
  res_degrees = 30/3600,
  parallel = FALSE,
  estimate_time = FALSE
)
}
\arguments{
\item{pops}{list of pop rasters}

\item{res_degrees}{resolution in degrees}

\item{parallel}{run parallel}

\item{estimate_time}{estimate time?}
}
\value{

}
\description{
Title
}
