% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{make_template}
\alias{make_template}
\title{Title}
\usage{
make_template(pops, res_degrees)
}
\arguments{
\item{pops}{list of pop rasters}

\item{res_degrees}{resolution in degrees}
}
\value{
a raster
}
\description{
Title
}
\keyword{(just}
\keyword{To}
\keyword{case)}
\keyword{check}
\keyword{crs}
\keyword{do:}
\keyword{here}
\keyword{in}
\keyword{internal}
\keyword{too?}
